+++ libs/utils/at.c
@@ -1,8 +1,8 @@
 #include <string.h>
 
-size_t char_at(const char *data, const char ch) {
+long char_at(const char *data, const char ch) {
 	const size_t size = strlen(data);
-	size_t result = -1;
+	long result = -1;
 
 	for (size_t i = 0; i < size; ++i) {
 		if (data[i] == ch) {
